pyproject.toml
[project]
name = "ocr-pipe-test"
version = "1.0.0"
description = "Standardized OCR Pipeline with Docling and Tesseract integration"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    # Core ML/AI Dependencies (from Docling)
    "torch>=2.2.2,<3.0.0",
    "torchvision>=0,<1",
    "transformers>=4.42.0,<5.0.0",
    "accelerate>=1.2.1,<2.0.0",
    "safetensors[torch]>=0.4.3,<1",
    "huggingface_hub>=0.23,<1",
    
    # Image Processing
    "Pillow>=10.0.0,<12.0.0",
    "opencv-python-headless>=4.6.0.66,<5.0.0.0",
    "numpy>=1.24.4,<3.0.0",
    
    # Data Processing
    "jsonlines>=3.1.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pandas>=1.5.0,<3.0.0",
    
    # OCR Dependencies (Tesseract)
    "pytesseract>=0.3.10,<1.0.0",
    
    # Utilities
    "tqdm>=4.64.0,<5.0.0",
    "rtree>=1.0.0",
    
    # Document Processing
    "docling-core>=2.19.0,<3.0.0",
    
    # Additional utilities for reconstruction
    "reportlab>=4.0.0,<5.0.0",
    "matplotlib>=3.5.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "black>=24.4.0,<25.0.0",
    "isort>=5.10.0,<6.0.0",
    "mypy>=1.10.0,<2.0.0",
    "flake8>=7.1.0,<8.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0,<9.0.0",
    "black>=24.4.0,<25.0.0",
    "isort>=5.10.0,<6.0.0",
    "mypy>=1.10.0,<2.0.0",
    "flake8>=7.1.0,<8.0.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
py_version = 39
